requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
python-jose[cryptography]
bcrypt==3.2.2
python-multipart
pydantic[email]